+++ test/cupertino_tab_test.dart
@@ -56,7 +56,7 @@ void main() {
     );
 
     final pageState = CupertinoTabPage.of(pageKey.currentContext!);
-    final tabController = pageState.tabController;
+    final tabController = pageState.controller;
 
     expect(pageState.index, 0);
     expect(tabController.index, 0);
@@ -99,7 +99,7 @@ class HomePage extends StatelessWidget {
     final tabState = CupertinoTabPage.of(context);
 
     return CupertinoTabScaffold(
-      controller: tabState.tabController,
+      controller: tabState.controller,
       tabBuilder: tabState.tabBuilder,
       tabBar: CupertinoTabBar(
         items: [